cleanup creation of datasources in storage provider tests
