+++ core/src/test/java/org/jobrunr/storage/sql/db2/AbstractDB2StorageProviderTest.java
@@ -1,16 +1,20 @@
 package org.jobrunr.storage.sql.db2;
 
+import com.zaxxer.hikari.HikariDataSource;
 import org.jobrunr.storage.sql.SqlStorageProviderTest;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.extension.AfterAllSubclasses;
 import org.junit.jupiter.extension.BeforeAllSubclasses;
 import org.junit.jupiter.extension.ForAllSubclassesExtension;
 import org.testcontainers.containers.Db2Container;
 
+import javax.sql.DataSource;
 import java.time.Duration;
 import java.time.Instant;
 
 import static java.time.Instant.now;
+import static org.jobrunr.storage.sql.SqlTestUtils.toHikariDataSource;
 
 @ExtendWith(ForAllSubclassesExtension.class)
 public abstract class AbstractDB2StorageProviderTest extends SqlStorageProviderTest {
@@ -18,6 +22,22 @@ public abstract class AbstractDB2StorageProviderTest extends SqlStorageProviderT
     protected static final Db2Container sqlContainer = new Db2Container("icr.io/db2_community/db2:12.1.0.0")
             .withPrivilegedMode(true)
             .acceptLicense();
+
+    protected static HikariDataSource dataSource;
+
+    @Override
+    public DataSource getDataSource() {
+        if (dataSource == null) {
+            dataSource = toHikariDataSource(sqlContainer);
+        }
+        return dataSource;
+    }
+
+    @AfterAll
+    public static void destroyDatasource() {
+        dataSource.close();
+        dataSource = null;
+    }
 
     @BeforeAllSubclasses
     public static void startSqlContainer() {
