+++ core/src/test/java/org/jobrunr/storage/sql/db2/DB2TablePrefixStorageProviderTest.java
@@ -24,22 +24,6 @@ import static org.jobrunr.utils.resilience.RateLimiter.Builder.rateLimit;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class DB2TablePrefixStorageProviderTest extends AbstractDB2StorageProviderTest {
 
-    private static HikariDataSource dataSource;
-
-    @Override
-    protected DataSource getDataSource() {
-        if (dataSource == null) {
-            dataSource = toHikariDataSource(sqlContainer);
-        }
-        return dataSource;
-    }
-
-    @AfterAll
-    public static void destroyDatasource() {
-        dataSource.close();
-        dataSource = null;
-    }
-
     @BeforeAll
     void runInitScript() {
         doInTransaction(getDataSource(),
