+++ core/src/test/java/org/jobrunr/storage/sql/h2/CommonsDbcpAutoCommitFalseH2StorageProviderTest.java
@@ -5,7 +5,7 @@ import org.apache.commons.dbcp2.BasicDataSource;
 public class CommonsDbcpAutoCommitFalseH2StorageProviderTest extends CommonsDbcpH2StorageProviderTest {
 
     @Override
-    protected BasicDataSource getDataSource() {
+    public BasicDataSource getDataSource() {
         return getDataSource(false);
     }
 }
