+++ core/src/test/java/org/jobrunr/storage/sql/h2/NativePoolH2TablePrefixStorageProviderTest.java
@@ -36,7 +36,7 @@ public class NativePoolH2TablePrefixStorageProviderTest extends SqlStorageProvid
     }
 
     @Override
-    protected DataSource getDataSource() {
+    public DataSource getDataSource() {
         if (dataSource == null) {
             dataSource = JdbcConnectionPool.create("jdbc:h2:mem:test-native-pool;INIT=CREATE SCHEMA IF NOT EXISTS SOME_SCHEMA;DB_CLOSE_DELAY=-1;", "sa", "sa");
         }
