+++ core/src/test/java/org/jobrunr/storage/sql/h2/TomcatJdbcPoolAutoCommitFalseH2StorageProviderTest.java
@@ -5,7 +5,7 @@ import org.apache.tomcat.jdbc.pool.DataSource;
 public class TomcatJdbcPoolAutoCommitFalseH2StorageProviderTest extends TomcatJdbcPoolH2StorageProviderTest {
 
     @Override
-    protected DataSource getDataSource() {
+    public DataSource getDataSource() {
         return getDataSource(false);
     }
 }
