+++ core/src/test/java/org/jobrunr/storage/sql/mariadb/MariaDbStorageProviderTest.java
@@ -9,19 +9,4 @@ import static org.jobrunr.storage.sql.SqlTestUtils.toHikariDataSource;
 
 class MariaDbStorageProviderTest extends AbstractMariaDbStorageProviderTest {
 
-    private static HikariDataSource dataSource;
-
-    @Override
-    protected DataSource getDataSource() {
-        if (dataSource == null) {
-            dataSource = toHikariDataSource(sqlContainer, "?rewriteBatchedStatements=true&useBulkStmts=false");
-        }
-        return dataSource;
-    }
-
-    @AfterAll
-    public static void destroyDatasource() {
-        dataSource.close();
-        dataSource = null;
-    }
 }
