+++ core/src/test/java/org/jobrunr/storage/sql/mysql/AbstractMySQLStorageProviderTest.java
@@ -1,21 +1,41 @@
 package org.jobrunr.storage.sql.mysql;
 
+import com.zaxxer.hikari.HikariDataSource;
 import org.jobrunr.storage.sql.SqlStorageProviderTest;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.extension.AfterAllSubclasses;
 import org.junit.jupiter.extension.BeforeAllSubclasses;
 import org.junit.jupiter.extension.ForAllSubclassesExtension;
 import org.testcontainers.containers.MySQLContainer;
 
+import javax.sql.DataSource;
 import java.time.Duration;
 import java.time.Instant;
 
 import static java.time.Instant.now;
+import static org.jobrunr.storage.sql.SqlTestUtils.toHikariDataSource;
 
 @ExtendWith(ForAllSubclassesExtension.class)
 public abstract class AbstractMySQLStorageProviderTest extends SqlStorageProviderTest {
 
     protected static MySQLContainer sqlContainer = new MySQLContainer<>("mysql:8.0.32").withEnv("TZ", "UTC");
+
+    protected static HikariDataSource dataSource;
+
+    @Override
+    public DataSource getDataSource() {
+        if (dataSource == null) {
+            dataSource = toHikariDataSource(sqlContainer, "?rewriteBatchedStatements=true&useSSL=false");
+        }
+        return dataSource;
+    }
+
+    @AfterAll
+    public static void destroyDatasource() {
+        dataSource.close();
+        dataSource = null;
+    }
 
     @BeforeAllSubclasses
     public static void startSqlContainer() {
