+++ core/src/test/java/org/jobrunr/storage/sql/oracle/OracleTablePrefixStorageProviderTest.java
@@ -28,29 +28,6 @@ import static org.jobrunr.utils.resilience.RateLimiter.Builder.rateLimit;
 @RunTestIfDockerImageExists("container-registry.oracle.com/database/standard:12.1.0.2")
 public class OracleTablePrefixStorageProviderTest extends AbstractOracleStorageProviderTest {
 
-    private static HikariDataSource dataSource;
-
-    @Override
-    protected DataSource getDataSource() {
-        if (dataSource == null) {
-            // dataSource = toHikariDataSource("jdbc:oracle:thin:@localhost:1527:xe".replace(":xe", ":ORCL"), "system", "oracle");
-
-            System.out.println("==========================================================================================");
-            System.out.println(sqlContainer.getLogs());
-            System.out.println("==========================================================================================");
-
-            dataSource = toHikariDataSource(sqlContainer.getJdbcUrl().replace(":xe", ":ORCL"), sqlContainer.getUsername(), sqlContainer.getPassword());
-        }
-
-        return dataSource;
-    }
-
-    @AfterAll
-    public static void destroyDatasource() {
-        dataSource.close();
-        dataSource = null;
-    }
-
     @BeforeAll
     void runInitScript() {
         doInTransaction(getDataSource(), statement -> {
