+++ core/src/test/java/org/jobrunr/storage/sql/postgres/PostgresStorageProviderTest.java
@@ -9,19 +9,4 @@ import static org.jobrunr.storage.sql.SqlTestUtils.toHikariDataSource;
 
 class PostgresStorageProviderTest extends AbstractPostgresStorageProviderTest {
 
-    private static HikariDataSource dataSource;
-
-    @Override
-    protected DataSource getDataSource() {
-        if (dataSource == null) {
-            dataSource = toHikariDataSource(sqlContainer);
-        }
-        return dataSource;
-    }
-
-    @AfterAll
-    public static void destroyDatasource() {
-        dataSource.close();
-        dataSource = null;
-    }
 }
